Fixed most restructuring problems
