+++ game/application/pom.xml
@@ -19,6 +19,11 @@
         </dependency>
         <dependency>
             <groupId>de.gleex.pltcmd</groupId>
+            <artifactId>game.options</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>de.gleex.pltcmd</groupId>
             <artifactId>model.radio</artifactId>
             <version>${project.version}</version>
         </dependency>
