+++ game/options/src/main/kotlin/de/gleex/pltcmd/options/GameOptions.kt
@@ -1,7 +1,5 @@
 package de.gleex.pltcmd.options
 
-import de.gleex.pltcmd.model.radio.AttenuationModel
-import de.gleex.pltcmd.model.radio.PercentageReducingAttenuation
 import de.gleex.pltcmd.model.world.Sector
 import org.hexworks.cobalt.databinding.api.extension.createPropertyFrom
 import org.hexworks.cobalt.databinding.api.property.Property
@@ -11,8 +9,6 @@ import org.hexworks.zircon.api.data.Size3D
  * Options that change the behaviour of the game.
  */
 object GameOptions {
-    // FIXME results in a dependency loop!
-    val attenuationModel: Property<AttenuationModel> = createPropertyFrom(PercentageReducingAttenuation())
 
     /**
      * Option to turn on the RadioSignalVisualizer and corresponding UI elements.
