+++ model/radio/src/test/kotlin/de/gleex/pltcmd/model/radio/RadioSignalTest.kt
@@ -1,9 +1,9 @@
 package de.gleex.pltcmd.model.radio
 
+import de.gleex.pltcmd.model.radio.testhelpers.shouldBeExactly
 import de.gleex.pltcmd.model.world.terrain.Terrain
 import de.gleex.pltcmd.model.world.terrain.TerrainHeight
 import de.gleex.pltcmd.model.world.terrain.TerrainType
-import de.gleex.pltcmd.testhelpers.shouldBeExactly
 import io.kotlintest.matchers.doubles.plusOrMinus
 import io.kotlintest.shouldBe
 import io.kotlintest.specs.WordSpec
