+++ pom.xml
@@ -15,7 +15,8 @@
 	<modules>
 		<module>model</module>
 		<module>game</module>
-	</modules>
+        <module>util</module>
+    </modules>
 
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
