--- src/test/resources/logback-test.xml
@@ -1,18 +1,0 @@
-<configuration>
-
-    <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-        <encoder>
-            <pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
-        </encoder>
-    </appender>
-
-    <root level="info">
-        <appender-ref ref="STDOUT"/>
-    </root>
-
-    <logger name="org.hexworks.zircon" level="INFO"/>
-    <logger name="org.hexworks.cobalt" level="INFO"/>
-
-    <logger name="de.gleex.pltcmd" level="debug"/>
-
-</configuration>
