+++ util/pom.xml
@@ -3,20 +3,17 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>game</artifactId>
-        <groupId>de.gleex.pltcmd</groupId>
+        <artifactId>pltcmd</artifactId>
+        <groupId>de.gleex</groupId>
         <version>0.0.2-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>game.options</artifactId>
-
-    <dependencies>
-        <dependency>
-            <groupId>de.gleex.pltcmd</groupId>
-            <artifactId>model.radio</artifactId>
-            <version>0.0.2-SNAPSHOT</version>
-        </dependency>
-    </dependencies>
+    <groupId>de.gleex.pltcmd</groupId>
+    <artifactId>util</artifactId>
+    <packaging>pom</packaging>
+    <modules>
+        <module>tests</module>
+    </modules>
 
 </project>
