+++ util/tests/pom.xml
@@ -3,21 +3,13 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>model</artifactId>
+        <artifactId>util</artifactId>
         <groupId>de.gleex.pltcmd</groupId>
         <version>0.0.2-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>model.world</artifactId>
+    <artifactId>util.tests</artifactId>
 
-    <dependencies>
-        <dependency>
-            <groupId>de.gleex.pltcmd</groupId>
-            <artifactId>util.tests</artifactId>
-            <version>${project.version}</version>
-            <scope>test</scope>
-        </dependency>
-    </dependencies>
 
 </project>
