+++ util/tests/src/test/kotlin/io/kotlintest/provided/ProjectConfig.kt
